+++ sources/macos/Tenniarb/Tenniarb/model/ElementModel.swift
@@ -327,6 +327,9 @@ public class DiagramItem {
     var description: String? = nil
     
     var properties: ModelProperties = ModelProperties() // Extra nodes not supported directly by model.
+
+    // Images cache.
+    var images: [String:NSImage] = [:]
         
     var x: CGFloat = 0
     var y: CGFloat = 0