+++ sources/macos/Tenniarb/docs/Changelog.md
@@ -1,6 +1,6 @@
 #  Changelog
 
-## Version 1.0.1
+## Version 1.1
 
 * Element deletion was not used Undo/Redo.
 * Support Copy/Paste on element tree outline.
@@ -8,4 +8,6 @@
 * Fix floating point number rounding.
 * Fix link doubling when do duplicate.
 * Improve parsing performance.
+* Embedded images
+* Markdown highlighting.